:green_heart: Stale detection action workflow
Re-added the stale detection action workflow that will run once a day. With the growing number of issue and PRs we need to keep on top of them and make sure we don't let them hang around forever or be forgotten about.

For some reason the stale.yml file was deleted in one of the merges so adding it back. The state duration has been increased to 6 months!