Bug #387 Fix graph/mouse offset issue on MacOS (#471)
On MacOS the render target for GLCanvas is being scaled correctly (possibly because of custom code in JOGL but haven't confirmed this, however the mouse positions from Java AWT are coming through as not scaled (ie. MouseEvent.java ->event.getPosition()). This change adds code to get the correct scale from the GLCanvas' transform matrix and apply it to the mouse points.