:apple: Fix labels not rendering on MacOS (#486) (#490)
There is an issue on MacOS where JOGL code to change the GL context using release()/makeCurrent() can cause a thread lock. It looks like this issue was occurring when trying to update the glyph textures (nodeLabel and connectionLabel text). As the Blaze render calls happen after the labels it also caused the Blazes to disappear.  An explanation of the issue was found in Chapter 2 of 'Pro Java 6 3D Game Development':
   This coding approach means that the context is current for the
   entire duration of the thread's execution. This causes no problems
   on most platforms (e.g. it's fine on Windows), but unfortunately
   there's an issue when using X11. On X11 platforms, a AWT lock is
   created between the GLContext.makeCurrent() and GLContext.release()
   calls, stopping mouse and keyboard input from being processed.