Bugfix/scene batcher issue591 (#594)
* :art: Moved SceneBatchNGTest to utilities/LabelUtilitiesNGTest
Moved SceneBatchNGTest to utilities/LabelUtilitiesNGTest, tests are
checking functionality of LabelUtilities and not SceneBatcher.

* :art: FIxed argument order in assertEquals calls.

* :bug: splitTextIntoLines now splits on spaces if able.
Previously prevSpace was set to Integer.MAX_VALUE causing
reamining.indexOf(' ',prevSpace) to always return -1. I have changed it
to MIN_VALUE to stop this happening.
This meant that the logic for determining the value of trimPoint had to
be changed to account for the fact that by default prevSpace would be 
smaller than MAX_LINE_LENGTH_PER_ATTRIBUTE.
I also used this opportunity to break out the logic into an if else
statement for ease of reading.