:bug: Fixed exportGlyphTextures method. (#641)
Removal of potentially superfluous filetype extension was previously
handled manually. The previous method results in the file being saved
to the wrong location if the file path has a full stop but there is no
extension given. Method now uses a built in method to remove any
extension, fixing the problem.