:bug::art: Fixed infinite twin bug. Removed padding. Fixed a bug where it was possible for a vertex to get caught between two other verticies which each nudge it to within twin distance of the other without themselves ever being nudged. This occured when one of the verticies was in a higher level node of the QuadTree and hence did not show up as a possible collision for the subject being nudged whilst still having the nudged vertex appear as a possible. Reverted to moving both verticies to combat this.
Now instead of moving all collided verticies it only moves one, allowing
for the fact that the subject now gets moved when the first twin is nudged.

Removed padding as I don't intend to provide it as an option to users
and testing has revealed that a value of 0 is fine.