Feature/issue364 improve uncollide performance (#758)
* :fire: Deleting vestigial classes

* :arrow_up: Adding AssertJ dependency.
AssertJ is a library that provides more powerful assertion test than
those included in junit or testNG. In particular it allows for the
comarison of class instances without a bespoke equals method.

* :art: Create Dimensions enumerator
Allows for the restriction of the number of dimensions to 2 or 3 / 
those implemented.

* :art: Create ABstract Bounding Box
Created an AbstractBoundingBox class which bounding boxes of all 
dimensions must extend. Reduces code duplication and when combined
with the factory allows for further code reduction.

* :art: Implement 2d and 3d bounding boxes.
These boxes store details of the used boundaries of the graph. They
also allow for the selection of subboxes of the graph based on there
being 2^dimensions equal size cuboid boxes.

* :art: Implement AbstractTree
Created an abstract tree to reduce code duplication between trees of
different dimensions. TreeFactory allows for even further reductions
in code duplication and the abstraction of the number of dimensions
away from some code.

* :art: Implement 2D and 3D trees.
These tress are designed to split the graph into 2^dimensions equal
sized areas for faster collision detection. Nodes can only collide if
they are contained within the same subsection of the graph.
This split can occur multiple times based on the set depth or maximum
number of object in one area.

* :art: :racehorse: Created new uncollidePlugin
The new uncollide plugin is designed to be faster and maintain more
of the graphs structure than the original uncollide plugin.
The back end of the plugin has been redesigned and the plugin uses a
new algorithm.
The new algorithm separates "twin" nodes to a minimum distance 
(currently set by the user) and then expands the graph until there are
no collisions. This allows for a trade off between node density and 
speed/ maintaining graph structure.
This will originally be an experimental plugin to allow for feedback
and potential changes to the behaviour and arguments exposed.

* :art: Renamed two &three in enum to TWO & THREE respectively
Renamed to match naming conventions.

* :art: :bug: Renamed variables to avoid hiding other variables.
Renamed to avoid using the same name.

* :art: :bug: Added private constructor to override implicit public con.
From SonarQube:
Utility classes, which are collections of static members, are not meant to be instantiated. Even abstract utility classes, which can be extended, should not have public constructors.

Java adds an implicit public constructor to every class which does not define at least one explicitly. Hence, at least one non-public constructor should be defined.

* :art: :fire: Deleted unused code.

* :art: Added constructor to hide implicit public constructor.
From SonarQube:
Utility classes, which are collections of static members, are not meant to be instantiated. Even abstract utility classes, which can be extended, should not have public constructors.

Java adds an implicit public constructor to every class which does not define at least one explicitly. Hence, at least one non-public constructor should be defined.

* :art: :fire: Removed unused code.

* :art: Changed constructor to be private, public by mistake.

* :art: Fixed minor code smells.

* :art: Fixed minor code smells.

* :art: Changed default value of max expansions to 20.
20 is a reasonable number that works quite well on a large range of sizes.

* :art: Combined nudge2D and nudge3D
Combined the nudge2D and nudge3D methods into a single nudge method.
This reduces code duplication.
Added private nudge method to further reduce duplication.

* :art: :White_check_mark: Improved uncollide tests.
Added another node to both uncollide arrangement tests so that the
scenario where two nodes are in exactly the same location is handled.

* :art: Changed all members of deltas==0 condition to delta = 0
Changed the argument in an if statement from testing that every member
of deltas is 0 to checking that the value delta is 0. These are
functionally the same but this is much more readable.

* :art: :fire: Removed unused imports

* :art: Improved comments.

* :fire: Removed un-necessary toString methods.

* :art: Renamed test to match test naming convention

* :white_check_mark: testGetIndex not tests for all index values.
This will improve test coverage by reaching previously untested parts
of an if statement.

* :white_check_mark: testGetIndex now tests for all index values.
This will improve test coverage by reaching previously untested parts
of an if statement.

* :art: Made dimensions package private
Dimensions does not need to be public as no classes outside the package
are expected top use it.
If other classes wanted to use it then this enum would need to be moved
to a more central location.

* :white_check_mark: Added test for Dimensions
Added test for the getOptions method in Dimensions.

* :art: Made classes package/default access, constructors protected.

* :art: Made variables final where possible.

* :art: Made constructor protected instead of default.
As suggested in review.

* :art: Made constructor protected instead of default.
As suggested in review.

* :art: Made Dimensions enum public
Needs to be public because it is an argument for the ArrangementPlugin
which should remain public so that it can be registered as an 
arrangement plugin and used in other places.

* :art: Made create method protected.

* :art: Changes to access modifiers
Made get___Quadrant/Octant methods protected.
Constructor using given boundaries has been made private. It should not
be used outside of the class, a bounding box for a graphs makes no
sense without a graph.
Other constructor has been made protected.
Deliberately left class at default access level.

* :art: Changed protected to default in default access classes
After further discussion decided that the use of the protected access
modifier in package private/default access classes is incorrect
(despite the fact it looks neater) as the extra use case of being able 
to access the method from outside the package in a subclass can not be
achieved in a default access class.

* :art: Changed protected member variables to default in default access classes
After further discussion decided that the use of the protected access
modifier in package private/default access classes is incorrect
(despite the fact it looks neater) as the extra use case of being able 
to access the variable from outside the package in a subclass can not be
achieved in a default access class.