:art: :racehorse: Added refreshVisualProcessor call
Added refreshVisualProcessor call to perform initial render of graph
to the GLCanavas. This is required now that visualUpdates will not
automatically be triggered when opening a new graph. Testing shows this
to be more performant. This makes sense as now it is performing the
minimal number of visual updates that will render the graph to the 
screen whereas perviously it would do more than required (duplicates).