Performance/838 renderer speed (#848)
* :art: :racehorse: Added updateModCounts method

Added method updateModCounts to allow for the modification counters to
be updated without triggering visual updates. This will be used when
opening a graph for the first time top avoid unnecessary visual updates.
Generalised other update methods to allow for a graph argument to
facilitate this.

* :art: :racehorse: Update modcounts of visualAccess when opening graph

Now updates the modcounts of the GraphVisualAccess when constructing a 
visualManager. This allows for the updateFromIndigenousChanges listener
to be added without triggering every single visualUpdate.

* :art: :racehorse: Added refreshVisualProcessor call

Added refreshVisualProcessor call to perform initial render of graph
to the GLCanavas. This is required now that visualUpdates will not
automatically be triggered when opening a new graph. Testing shows this
to be more performant. This makes sense as now it is performing the
minimal number of visual updates that will render the graph to the 
screen whereas perviously it would do more than required (duplicates).

* :art: Deleted accidental blank line.

* :art: Made variables final where appropriate as per feedback.

* :racehorse: Moved rebuild from dispose to reshape.

Moved rebuild from dispose method to reshape method. This will stop a
rebuild from being triggered when a graph is closed and mean we can
remove the refresh from VisualGraphTopCOmponent as opening a graph
calls reshape (but not dispose).

* :racehorse: Moved rebuild from reshape to init.

Moved rebuild from reshape method to init method it must be called 
after a dispose to render the graph back to the screen (where the graph
hasn't actually been closed). An init must always be called in this
scenario.