:bug: Make addRectImage threadsafe
Made addRectImage more threadsafe by combining check to see if hashCode is contained in memory and the action to add the hadCode to the memory if it isnt. This elimnated multithreading scenario where the same hashcode is added twice ( next check for the presence of the hashcode is completed before it previous check finished adding it to memory) and the ptr values becomes incorrect.