:Art: Changed thread pool size to number of cores.
All three multithreaded methods are entirely CPU dependent. In this scenario the accepted wisdom is to have a number of threads equal to the number of cores. Setting each to the number of cores isn't perfect (as often more than one of these methods will run, such as when opening a graph. There are also potentially threads already running.) however accurately acounting for each scenario is quite involved and as far as I can tell there should be minimal penalty for having a few too many threads.