:racehorse: Streamline getMixedVertexCoordinates, setVertexCoordinates
Created new versions of getMixedVertexCoordinates and setVertexCoordinates which required the attributeIds that would normally be derived within the function to be provided as arguments. This allows the methods to run very slightly quicker. They should be used when the methods are expected to be called repeatedly with the same attribute Id's so that the ID's can be retrieved only once instead of every time they are run.