Performance/878 drag performance (#890)
* :art: :racehorse: Added updateModCounts method

Added method updateModCounts to allow for the modification counters to
be updated without triggering visual updates. This will be used when
opening a graph for the first time top avoid unnecessary visual updates.
Generalised other update methods to allow for a graph argument to
facilitate this.

* :art: :racehorse: Update modcounts of visualAccess when opening graph

Now updates the modcounts of the GraphVisualAccess when constructing a 
visualManager. This allows for the updateFromIndigenousChanges listener
to be added without triggering every single visualUpdate.

* :art: :racehorse: Added refreshVisualProcessor call

Added refreshVisualProcessor call to perform initial render of graph
to the GLCanavas. This is required now that visualUpdates will not
automatically be triggered when opening a new graph. Testing shows this
to be more performant. This makes sense as now it is performing the
minimal number of visual updates that will render the graph to the 
screen whereas perviously it would do more than required (duplicates).

* :art: Deleted accidental blank line.

* :art: Made variables final where appropriate as per feedback.

* :racehorse: Moved rebuild from dispose to reshape.

Moved rebuild from dispose method to reshape method. This will stop a
rebuild from being triggered when a graph is closed and mean we can
remove the refresh from VisualGraphTopCOmponent as opening a graph
calls reshape (but not dispose).

* :racehorse: Moved rebuild from reshape to init.

Moved rebuild from reshape method to init method it must be called 
after a dispose to render the graph back to the screen (where the graph
hasn't actually been closed). An init must always be called in this
scenario.

* :bug: Make addRectImage threadsafe

Made addRectImage more threadsafe by combining check to see if hashCode is contained in memory and the action to add the hadCode to the memory if it isnt. This elimnated multithreading scenario where the same hashcode is added twice ( next check for the presence of the hashcode is completed before it previous check finished adding it to memory) and the ptr values becomes incorrect.

* :art: Created GlyphStreamContext and NodeGlyphStreamContext

Created a GlyphStreamContext object which holds the information required by a GlyphStream to perform the compulsory methods. This will more easily allow for consistent method signatures across types of GlyphStreams.

* :art: Created NodeGlyphStream

Created a NodeGlyphStream object which contains logig to adding glyphs to a glyphStream focussed on Nodes. This logic previously resided in NodeLabelBuffer. The creation of this object will allow for the glyphStream to be treated separately to the buffer.

* :art: Accomodate new NodeGlyphStream and NodeGlyphStreamContext classes

Changes to method signatures to add in new object types NodeGlyphStream and NodeGlyphStreamContext.
NOTE: This commit does not contain all changes required for this. However other files have additional changes which should be explained in the commit.

* :art: Accomodate new classes, make RenderTextAsLigatures threadsafe.

Changes to method signatures to add in new object types NodeGlyphStream and NodeGlyphStreamContext.
Eliminate the reliance on class object drawing in method renderTextAsLigatures. Combined with changes to the methods which renderTextAsLigatures calls this has the affect of making the method more threadsafe.

* :racehorse: Add multithreading to methods

Made methods createBatch, fillTopLabels and fillBottomLabels multithreaded. This will improve the performance of these methods enabling faster application of graph changes. In particular refreshing the graph should be much quicker, leading to faster graph opening times.
NOTE: The branch, including this commit and all previous should work, however there has been limited testing. It is likely that more changes will be made.

* :arte: Incorporate GlyphStreamContext class

Change to method signatures and small change to logic to incorporate the new GlyphStreamContextClass. Should have been added to previous commit.

* :bug: Stronger synchronized condition to stop rare rendering bug

Combined array specific locks into one stronger lock that ensures
addition of nodes to both currentFloats and currentInts prior to allowing access to either. This seems to stop a glitch where characters would occasionally not appear in the node label when opening a graph.

* :art: Make GlyphStreamContext attributes public, remove get/set

Made GlyphStreamContext and child member variables public to improve
readability and ease of use.

* :racehorse: Make ConnectionLabelBatcher multithreaded.

Using the same parent classes and principles used when multithreading NodeLabelBatcher I have multithreaded ConnectionGlyphStream. This improved the rendering time of graphs dominated by connection count rather than node count.

* :racehorse: Multithread updateTop/BottomLabels

updateTopLabels and updateBottomLabels now use the multithreaded version of
FillTopLabels and FillBottomLabels respectively, brining them in line with changes madde to createBatch.

* :Art: Removing TODO comment (tried it)

Tried refactoring NodeLabelBatcher and NodeGlyphStream to run multiple
nodeGlyphStreams independently and then combine them to eliminate the need to synchronize on addLock. There was no improvement in speed, however my implementation likely wasn't optimal.

* :Art: Removing TODO comments (Done)

* :Art: Changed thread pool size to number of cores.

All three multithreaded methods are entirely CPU dependent. In this scenario the accepted wisdom is to have a number of threads equal to the number of cores. Setting each to the number of cores isn't perfect (as often more than one of these methods will run, such as when opening a graph. There are also potentially threads already running.) however accurately acounting for each scenario is quite involved and as far as I can tell there should be minimal penalty for having a few too many threads.

* :racehorse: scheduleXYZChangeOperation argument change.

Changed the argument used by DefaultInteractionEventHandler.scheduleXyzChangeOperation to an integer array of the verticies that have been moved. This changes processing downstream from processing every node to just those moved.

* :racehorse: Use vertexStream to improve performance

gatherSelectedNodes now uses a vertexStream to more quickly find selected nodes. This leads to better performance when dragging nodes.

* :racehorse: Streamline getMixedVertexCoordinates, setVertexCoordinates

Created new versions of getMixedVertexCoordinates and setVertexCoordinates which required the attributeIds that would normally be derived within the function to be provided as arguments. This allows the methods to run very slightly quicker. They should be used when the methods are expected to be called repeatedly with the same attribute Id's so that the ID's can be retrieved only once instead of every time they are run.

* Revert ":racehorse: Streamline getMixedVertexCoordinates, setVertexCoordinates"

This reverts commit a03581e5b7d0b09d1046338b6ceaa7ab44c0d7f1.

* :art: Refactored to be same as master.

I ruined the previous merge of master onto this branch. This corrects that.

* :art: Recommit changes

* :art: Remove unneccesary null check.
Removed code smell caused by unneccesary null check.