:racehorse:WIP Reduced memory usage and slightly improved performance.
Changed data structure for storing neighbour weights from an int[] to a HashMap. In the common cases where most verticies are not neighbours this will reduce memory usage by not strong 0s for all the vertex pairs without a transaction between them.
Vertex details are now stored in an object with neighbourWeight details to break up logic and stop the storage of every vertices weight together.
Next Step: Improve performance of adding results to the graph.