:bug: updateSelectablePluginsParameters() now updates aggregator param.
It appears that whenever a change is made that could affect one of the analytic parameters all of the parameters are being recreated from scratch. When making a box selection on the graph the aggregator parameter was not being recreated, this change will now cause it to be created. This stops a bug causing an error to occur when someone selects an analytic, changes the selection on the graph and then changes the aggregation method.
Note: I'm not sure why it needs to be rebuilt. I'm sure there is a better more correct way to fix this but as this is a blocking bug I want to get it fixed. Given this bug and performance problems I think really the analytic view needs a complete overhaul. Ill add a ticket for this.