:art: Make tests more readable and resistant to unimportant changes
previously the tests relied upon knowing the transactionID each 
cosine similarity result would be assigned to. This is unimportant to the analyst
and obtuse when reading the code. Switching to rely upon the more meaningful identifier of the
transaction to figure out which results we are testing means that the test is now resistant to changes
that may alter the order cosine similarity scores are generated but not the actual results and
is easier to read for future developers.