:bug: :white_check_mark: Updates tests. Test were previously incorrect, they looked for values that were slightly wrong. These values included a bug where the pageRank distributed to each other node by a sink would be its pageRank divided by the total number of nodes rather than the number of nodes minus 1 (as it wont distributed pageRank to itself).
I have modified the test to only perform 2 iterations. This should still test that multiple iterations works but is also now a small enough test for values to be calculated by hand in order to verify the values are correct. This allows the values to be set based on the true value and not the values being produced by the analytic at the time of writing the test.