Fix to show all instances of searched text in conversation bubbles
Styling does not match older code where there was some padding between the content and the conversation bubble. The usage of Selectable Label has been replaced with TextFlow which may be the reason, will continue to investigate.

Searched text is now styled in yellow with a black shadow for contrast as the previous method of highlighting styling does not work with TextFlow regions.

A benefit of not using highlighting is it cannot be accidentally unhighlighted by a user.

For now I have commented out the older code.