:art: :racehorse: Refactored computeShortestPathsUndirected
Refactored coomputeShortestPathUndirected to follow a more object 
orientated approach. It is hoped that this will reduce memory usage by
creating only what is needed. Previously multiple bitsets were created,
each of which had a size of vertexCount x vertexCount. Many of the 
entries of which would just be 0 and essentially ignored.