:bug: Stop error triggered when switching graphs with find open
Previously an error could occur when switching between graphs because
a FindRules attribute and type would be modified without its attributes
being updated to reflect the new type.
The line added causes the attributes value to be corrected prior to
accessing the attributes in the method that the error occurs in.