:bug: Fixed error when plugin reporter view not also open
Removed the call to newActiveGraph in graphOpened as in certain
situations this would cause a new listener to be added whilst the
list of listeners was being iterated over, which causes a oncurrentModification
exception.
I expected the removal of the call to cause behaviour issues, however
son far testing has not revealed any unexpected behaviour.