:bug: Potential fix to bug causing glyphs to replace others. (#963)
* :bug: Potential fix to bug causing glyphs to replace others.

A bug has been reported that the wrong letter may appear in the place
of another. As it happens for all occurances of that letter and the glyph
is pefectly formed it appears to be an issue with the logical connection
of a glyph to its position.
On inspections I dont believe this part of code was threadsafe, in
particular memory.size() may be 1 less than it should be if the previous
glyph is still being added to the map. This would cause the second glyph
to overwrite the first, subsequent occurances of the first glyph would then
retrieve the second glyph, which would explain the behaviour detailed
above.
This should fix that by ensuring that a new glyph is not added until the
previous has been added.

* :art: Made some variables final.