:art: Moved getSelectedElements logic to VisualGraphUtilities
DefaultInteractionEventHandler had a helper method to grather the 
selected elements. I have moved this logic to VisualGraphUtilities to
allow other classes to call this method as this logic is regularly
used in Constellation.
At a later date time should be spent to ensure this method is as quick
as possible and being used throughout the application.