:racecar: refactored calculatePertinentChanges to avoid duplicates.
Currently if there are multiple changes with the same master change
type we will perform this visual change multiple times regardless of
whether the change is the same or whether any information in the change
is utilised.
In particular when a Full_REFRESH is complete the whole graph will be
visually refreshed regardless of the change itself.
When deleting nodes from the graph this actually leads to the graph
being refreshed twice.
This change will stop this from occuring, this will speed up deleting nodes
amongst other things.