Performance/28 pagerank memory (#894)
* :racehorse: Various changes to reduce pagerank memory usage
Now stores the amount of outward links a vertex has in the separate outcounts map instead of storing it a number of times equal to the amount of verticies it neighbours.
Tidied edge direction logic.
Deal with sinks (nodes with no outgoing edges) more succinctly.
Added progress tracker.
No longer build extra set of neighbours for each vertex. Looks like these wern't being garbage collected properly and may have been the main cause of the memory issue.
Moved maxPagerank and delta calculation to after pagerank calculation instead of during, this is to set up for trialing using the stream framework or parallelising the pagerank calculation to improve performance (neither of these have been done yet).

* :bug: Fixed bug causing early finish
Delta was being calculated incorrectly. This would cause the epsilon to be met prematurely and the analytic to finish prematurely.

* :racehorse: :art: Complete refactor of PagerankCentralityPlugin.edit
Created new subclass PagerankVertex to enable sotrage of related information together and allow quicker access to data such as pageRank, neighbours etc
This also reduces the amount of duplicate data and makes the code more readable through the creation of multiple new methods.

Improves both the speed and memory usage of the analytic.

* :art: Made variables final and private where applicable.

* :art: Updated author due to wholesale changes.

* :art: Fixed code smells.

* :art: Made changes requested in pull request.

* :bug: :white_check_mark: Updates tests.
Test were previously incorrect, they looked for values that were slightly wrong.
These values included a bug where the pageRank distributed to each other node by a sink would
be its pageRank divided by the total number of nodes rather than the number of nodes minus 1 (as it wont distributed pageRank to itself).

I have modified the test to only perform 2 iterations. This should still test that multiple iterations works but is also now a small enough test for values to be calculated by hand in order to verify the values are correct. This allows the values to be set based on the true value and not the values being produced by the analytic at the time of writing the test.

* :art: Seperation of controlling logic and PageRankVertex object.

The main change is the seperation of graph/analytic level actions with
PageRankVerticies from the actual PageRankVertex. This has been combined
with other smaller recommendations made in the pull request for this
branch.

* :art: Removed vestigial lines from testing