Performance/28 cosine (#908)
* :racehorse: Various changes to reduce pagerank memory usage
Now stores the amount of outward links a vertex has in the separate outcounts map instead of storing it a number of times equal to the amount of verticies it neighbours.
Tidied edge direction logic.
Deal with sinks (nodes with no outgoing edges) more succinctly.
Added progress tracker.
No longer build extra set of neighbours for each vertex. Looks like these wern't being garbage collected properly and may have been the main cause of the memory issue.
Moved maxPagerank and delta calculation to after pagerank calculation instead of during, this is to set up for trialing using the stream framework or parallelising the pagerank calculation to improve performance (neither of these have been done yet).

* :bug: Fixed bug causing early finish
Delta was being calculated incorrectly. This would cause the epsilon to be met prematurely and the analytic to finish prematurely.

* :racehorse: :art: Complete refactor of PagerankCentralityPlugin.edit
Created new subclass PagerankVertex to enable sotrage of related information together and allow quicker access to data such as pageRank, neighbours etc
This also reduces the amount of duplicate data and makes the code more readable through the creation of multiple new methods.

Improves both the speed and memory usage of the analytic.

* :art: Made variables final and private where applicable.

* :art: Updated author due to wholesale changes.

* :art: Fixed code smells.

* :art: Made changes requested in pull request.

* :racehorse:WIP Reduced memory usage and slightly improved performance.

Changed data structure for storing neighbour weights from an int[] to a HashMap. In the common cases where most verticies are not neighbours this will reduce memory usage by not strong 0s for all the vertex pairs without a transaction between them.
Vertex details are now stored in an object with neighbourWeight details to break up logic and stop the storage of every vertices weight together.
Next Step: Improve performance of adding results to the graph.

* :art: Separated attribute ensuring logic from score adding logic.

Created method setGraphAndEnsureAttributes to separate the logic for ensuring the graph has the right attributes from the logic that adds the scores to the graph. This is to pave the way for publically allowing scored to be added individually without repeatedly ensuring attributes are on the graph.

* :racehorse: Refactored addScoreToGraph to contain bulk of logic.

Refactored the addScoreToGraph method to contain the bulk of the logic so that addScoresToGraph now calls addScoreToGraph multiple times rather than addScoreToGraph creating a collection and calling addScoresToGraph.

Clarified some logic and deleted extrenuous comments in CosineSimilarityPlugin.

* :racehorse: Modified plugins to make use of addScoreToGraph method

Modified analytic plugins currently using the addScoreToGraph plugin to instead make use of the addScoreToGraph plugin.
The eliminated the need to gather all scores prior to adding them to the graph, reducing memory usage and decreesing the time taken to add scores to the graph.

* :art: Fixed typo

* :racehorse:Early removal of verticies unable to pass minCommonFeatures.

Instead of adding every vertex with atleast one neighbour to the list of vertices we now only add vertices with more neighbours than the minCommonFeatures values.
If minCommonFeatures > 1 then this will reduce the number of times that the common neighbours are calculated and increase the speed of the plugin.

* :art:Removed unusued parameter from addScoreToGraph.

* :art: Renamed variable to avoid hiding class variable.

* :art: Removed unused logger.

* :art: Made applicable variables and arguments final.

* :art: Removes unused method agrument.

* :art: Make tests more readable and resistant to unimportant changes

previously the tests relied upon knowing the transactionID each 
cosine similarity result would be assigned to. This is unimportant to the analyst
and obtuse when reading the code. Switching to rely upon the more meaningful identifier of the
transaction to figure out which results we are testing means that the test is now resistant to changes
that may alter the order cosine similarity scores are generated but not the actual results and
is easier to read for future developers.

* :bug: Fixed bug meaning only results > minCommonFeatures were calculated

Relaxed minCommonFeatures condition from > to >= to stop the analytic ignoring similarities
of interest.Also made a small change to connectionOn/Out logic so that it revolved around the vertex
of interest and not the neighbour, I think this is just slightly clearer.

* :bug: :white_check_mark: Updates tests.
Test were previously incorrect, they looked for values that were slightly wrong.
These values included a bug where the pageRank distributed to each other node by a sink would
be its pageRank divided by the total number of nodes rather than the number of nodes minus 1 (as it wont distributed pageRank to itself).

I have modified the test to only perform 2 iterations. This should still test that multiple iterations works but is also now a small enough test for values to be calculated by hand in order to verify the values are correct. This allows the values to be set based on the true value and not the values being produced by the analytic at the time of writing the test.

* :bug: :white_check_mark: Updated tests.
Updates tests to rely on the identifier of the transaction, not the transaction ID
as the transaction ID could change without the results being changed.

* :bug: :white_check_mark: Updated tests.
Updates tests to rely on the identifier of the transaction, not the transaction ID
as the transaction ID could change without the results being changed.

* :art: Minor improvements suggested in PR review.

* :art: Minor improvements suggested in PR review.

* Revert ":art: Minor improvements suggested in PR review."

This reverts commit 4fd734e5470f5a80b56e864fdf0ecc3738f4e011.

* :art: Small changed recommended in PR

Removed unneeded line
Prioritize primitve types where possible