Link notes to nodes and transactions
User Notes store which nodes and transactions were selected when the note was created.
Added ability for users to change the selection for a specific note from the edit dialog.
Right click menu added to user notes for users to select the related nodes and transactions on the graph for that note.
When in the selection filter, notes relating to the graph elements currently highlighted appear.