Modify shaders to draw greyscale for anaglyphic display.
Anaglyphic display relies on drawing two different colors, one for each eye.
This means that the usual graph colors should be discarded and everything
drawn in greyscale using a glColorMask to draw the two colors.

This commit modifies the batchers and shaders for nodes, lines, linelines,
loops, and blazes. A uniform variable is added and set when things are to
be drawn as grey using a simple "convert to grey" expression.

Selecting anaglyphic drawing is purely a visual thing, with no changes to the
graph. Therefore, you have to make a change to the graph (move/zoom/pan)
before the display changes.

The glasses colors (left eye green, right eye magenta) are still harcoded.