Fix null pointer exception
When Graph Element options that had Category: Graph Property, Property: Neighbour Count and Bin Into Buckets or Bin Into Logarithmic Buckets chosen, they threw a null pointer exception when switching “to” it.

This has been fixed by checked if the plugin parameters required are null, and if so calling createParameters().