Find View UI Improvements (#1734)
* Changed text in find dialog box

Changed title from "Find View" to "Find and Replace". Changed text in the "Look for" drop down to plural.

* Attribute option changed

The "In Attributes" text has been changed to "Search Attributes".  All attributes are selected by default.

* Advanced Pane Padding

Set padding between "Match Criteria" and "Current Selection". Changed "aa" to "Ignore Case" in the StringCriteriaPanel class.

* Fixed Issue With "Ignore Case" Label

Clicking on "Ignore Case" in the advanced panel changed it back to "Aa" by a function call. The call has been commented out.

* Advanced Search Displays Number of Results Found

* Simplified Implementation of Displaying Number of Results

* Updated Unit Tests

* Fixed Code Smells

* Fixed More Code Smells

There is still 1 left which is just adding more comments but the class is already well commented.

* Made Changes to Label Position

The label showing matches found is at the bottom now. It only appears when a search is made.

* Made Requested Changes

* Removed Some Code Smells

* Fixed Issue with Advanced Tab